Node.js项目-[67]-职位管理-编辑职位-3

+++ web-tutorial/2021年Nodejs教程/lagou-admin/backend/middlewares/upload.js
@@ -55,12 +55,12 @@ const uploadMiddleware = (req, res, next) => {
 				}),
 			})
 		} else {
-			if (filename !== "") {
-				/**
-				 * 前端上传新的图片，就在后端把旧的图片删掉
-				 * 新加的字段companyLogo_old不是在model定义的，所以没有显示在数据库
-				 */
-				const { companyLogo_old } = req.body
+			/**
+			 * 前端上传新的图片，就在后端把旧的图片删掉
+			 * 新加的字段companyLogo_old不是在model定义的，所以没有显示在数据库
+			 */
+			const { companyLogo_old } = req.body
+			if (filename !== "" && companyLogo_old) {
 				try {
 					fs.unlinkSync(
 						path.join(
@@ -68,12 +68,12 @@ const uploadMiddleware = (req, res, next) => {
 							`../public/uploads/${companyLogo_old}`
 						)
 					)
-					req.companyLogo = filename
 				} catch (err) {
 					console.log(err)
 				}
-				next()
 			}
+			req.companyLogo = filename
+			next()
 		}
 	})
 }
